実行系の認可と Engine hydrate を AuthorizationGuard / EngineSession に集約し、横断関心の重複を削減する。
