+++ core/application/Statevia.Core.Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -24,6 +24,8 @@ public static class ApplicationServiceCollectionExtensions
         services.AddSingleton<IActionSchemaService, ActionSchemaService>();
         services.AddSingleton<IDefinitionSchemaService, DefinitionSchemaService>();
         services.AddSingleton<ExecutionOwnershipTracker>();
+        services.AddScoped<ExecutionAuthorizationGuard>();
+        services.AddScoped<ExecutionEngineSession>();
         services.AddScoped<ExecutionQueryService>();
         services.AddScoped<ExecutionIdempotencyService>();
         services.AddScoped<ExecutionProjectionOrchestrator>();
